requirements.txt
openai>=1.0.0
aiohttp>=3.8.0
asyncio>=3.4.3
python-dotenv>=1.0.0
tiktoken>=0.5.0
